apps/websocket-gateway/requirements.txt
# WebSocket Gateway Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
nats-py==2.6.1
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
python-json-logger==2.0.7
prometheus-client==0.19.0